API Version Locking Implementation
Version locking is a mechanism for preventing the system from running an outdated api version. This prevents incidences of bugs recurring, possible crashes, data integrity issues introduced in previous API versions.

This feature can be enabled or disabled in the configuration file but is enabled by default without any config.

MasterConfig.ts is where you can configure API version locking. This is the preferered configuration file because it cant be easily modified once the application is built unlike config.json