TX TB workflow
- Can now enter First tb treatment date and TB treatment period

Note: On followup visit depending if the client is still on treatment, we dont ask them to re-enter treatment date and period. This question is asked again after treatment period is over.

Also note that concept for "TB treatment period" uses a placeholder concept_id because the concept doent exist at the time of this commit. @brian-mw and @Skipper-116 take note.